+++ src/App.tsx
@@ -1,7 +1,7 @@
 import Button from './components/atoms/button/Button';
 import React from 'react';
 import logo from './logo.svg';
-import styles from './styles.module.scss';
+import styles from './style.module.scss';
 
 function App() {
   return (
